Add CIJ_transform_M and convert CIJ_rot3 to use it
CIJ_transform_M (anisotropy_ajn.f90) uses a different method to
transform an elastic tensor which permits one to directly manipulate
the 6x6 Voigt matrix representation.  Consequently, there are fewer
operations and the procedure is faster.  It is based on the formulae
in A.F. Bower, 'Applied mechanics of solids', section 3.2.11, at
http://solidmechanics.org/Text/Chapter3_2/Chapter3_2.php#Sect3_2_11

CIJ_rot3 now makes use of CIJ_test_transform_M and is quicker than
before.

It can also be used to flip or rotate constants about an arbitrary
plane or axis, and may be useful to rotate with Euler angles in the
future.