spherical_geometry: Add routines to project points onto great circles
Also add a number of small utility functions.  This begins the
move to a consistent naming scheme: {function,subroutine} sg_*.
This helps avoid namespace issues inherent to Fortran.

The sg_project_to_gcp subroutine takes a pole to a great circle
and a point and projects that point onto the great circle along
the great circle containing the pole and the point (which is the
shortest way to get to the point).