anisotropy_ajn: Add CIJ_stable to test tensor stability
CIJ_stable checks that matrices are:
- symmetrical; and
- positive definite.
The latter is a requirement for the elasticity tensor to be stable, and
in the current implementation we use Sylvester's criterion to decide
(which states that all leading minors of the matrix must have a positive
determinant).  Add the function determinant() for this purpose.