f90sac: Remove C I/O and tidy source
Remove the need to link to C routines by using Fortran
stream access.  This is the same speed as C, so the
new routines will be the same speed as the old.

Existing codes compiled against the library are not
affected.

New, private routines remove code duplication for file
I/O:
 - f90sac_open_writeheader() opens a logical unit, reads
   the header into the structure, and leaves the unit
   open.  This is used by f90sac_write{header,trace}().
 - f90sac_open_readheader() does the same as above, but
   for reading.  It is called by f90sac_read{trace,header}().

Other changes to tidy source:
 - Make routines and variables private by default.
 - Re-order the list of public routines to be alphabetical.
 - Remove unused variables.