FFFTW: Update FFTW calls to use functions where needed
At some point, the modern F2003 FFTW interface has been changed
so that the equivalent C functions which return something have
become functions in Fortran as well.  For instance:
    plan = fftw_plan_dft_1d(n, in, out, FFTW_FORWARD, FFTW_ESTIMATE)
where `plan` is a type(C_PTR).

Previously, the returned thing was the first argument to a
subroutine, and this worked with older GCC versions.  GCC5
picks this up as an error.